+++ test/EndToEnd/RuleSet/Php83Test.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2019-2023 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/php-cs-fixer-config
+ */
+
+namespace Ergebnis\PhpCsFixer\Config\Test\EndToEnd\RuleSet;
+
+use PHPUnit\Framework;
+
+#[Framework\Attributes\CoversNothing()]
+#[Framework\Attributes\RequiresPhp('8.3')]
+final class Php83Test extends AbstractRuleSetTestCase
+{
+}